Save the channel a reading was received in into the db as well.
When we started having nodes like tinytemp with two tmep36 sensors, it started to _matter_ whether the reading was from
channel 1 or 2.  In the live graphing web pages, this was read out simply by the index of the sensor in the kpacket,
but when reading historical data out of the db, this information was lost!