Demo working frequency, adc and internal temp, with dual USB pipelines.
It doesn't seem particularly stable though.  If I connect to /dev/ttyACM1,
it seems to happily and reliably keep piping out the string formatted data,
but rx from the xbee, and piping to /dev/ttyACM0 seems to often get bad packets.

Perhaps we're not reading from the xbee often enough?