Fix frequency counting.
Original code on teensy humi was actually wrong, sizeof(unsigned int) is actually only 16 bits on avrs,
so what I was using as the "working reference" was actually only the lower 16 bits.

Likewise, the odd behaviour with version and sensor count is because ordering and packing of bit fields in C is
implementation dependent, and nothing to ever rely on.