Add a simple tx library routine.
Supports 16 bit addressing only, no acks, and only to the local PAN.

About the fudge factor:
Even though the mrf24j40 datasheet and the 802.15.4 standard seem to agree, my
xbee receivers never ever received the first two bytes of data, but all the
checksums were correct.  Assuming I had an error in my bit field calculation
for the variable length of the addressing bytes, I tried the following
combinations, with verification on the receiver, for a four byte packet,
p[0,1,2,3]

* 16bit src, no src pan, 16bit dest, dest pan, pan compression bit set:
  correct src, correct dest, received size 2, p[2,3]
* 16 bit src, src pan, 16bit dest, dest pan, no pan compression:
  same
* 64 bit dest, no src pan, 16bit dest, dest pan, pan compression:
  same, again, valid 64 bit src.

I tried setting the header length to +2, to see if the mrf24j40 module was
expecting the header length and frame length bytes to be included, but that
introduced errors.  Simply setting the frame length to +2, and skipping two
bytes after the headers produced clean checksum, "correct" length packets on
the receiver.