Support decoding of reports from the fridge controller.
Support the other endian of the mrf24j chip.
Add threshold and min runtime config for the relay output.  (Verified with a light!)

While the EEPROM config block works, I don't have the rf reception working yet.  The tiny84 is a bit
light on debugging tools (other than the rf link under test) so I think I'll wire this back up on a
teensy board and get some printout of what's going on....