Decode a lot more of the 802.15.4 headers and put them in the rx info object.
Note, the variable length decoding for addressing is still left to the client. The demo code has been expanded
to do basic decoding for the 16bit pan compressed addressing I normally use.

FCS bytes are no longer counted in the rxbuffer length, because, really, who cares.  The hardware has already
taken care of checking it.  If you _want_ the fcs bytes are actually hiding at the end of the rxbuffer, in the
two bytes after the frame length.

Now, the "packet data" bytes shown are _only_ payload bytes.