Totally new layout targetting C4 devices.
Only has SPI1, so can't use SPI2.  Revised layout to vaour vias
over through hole parts a) more compact and b) when cut on a milling
machine, it's easier to put in a via than try and make sure to solder
both sides of through hole parts.

Correctly provide power to both jacks and the RHT03 device from the mcu
itself, instead of leaving them permanently powered.
Correctly include a wakeup pin to the MRF to allow switching it off when desired.
Correctly place the bigger capacitor on the VDD3 pin as stated in the
data sheets.
Use total of 800k resistance instead of 1.5M for the resistor divider. A note
in the datasheet talks about problems with stray resistance on the board
interfering if you go to high, and I was seeing something like this.
Include more notes, including values for running at a higher voltage.
This seems to be critically important for the RHT03 sensor. At 3.0V,
it responds, but the waveform is more of a ripple at 3V than a 0-3v3 signal.